refactor auth wrapper into utils
