+++ frontend/utilities/device-token.ts
@@ -2,31 +2,6 @@ import * as Notifications from "expo-notifications";
 import { Platform } from "react-native";
 import * as Device from "expo-device";
 import Constants from "expo-constants";
-import AsyncStorage from "@react-native-async-storage/async-storage";
-
-// retrieve JWT token from local storage
-export const getAuthToken = async (): Promise<string | null> => {
-  try {
-    // Get all keys and find the Supabase auth token key
-    const allKeys = await AsyncStorage.getAllKeys();
-    const supabaseTokenKey = allKeys.find(
-      (key) => key.startsWith("sb-") && key.endsWith("-auth-token"),
-    );
-
-    if (!supabaseTokenKey) {
-      return null;
-    }
-
-    const tokenString = await AsyncStorage.getItem(supabaseTokenKey);
-    if (!tokenString) return null;
-
-    const parsedToken = JSON.parse(tokenString);
-    return parsedToken?.access_token ?? null;
-  } catch (error) {
-    console.error("Failed to get auth token:", error);
-    return null;
-  }
-};
 
 // get Expo push notification token
 export const getExpoDeviceToken = async (): Promise<string | null> => {
